requirements.txt
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-cors-headers==2.5.2
django-phonenumber-field==2.3.1
djangorestframework==3.9.2
djangorestframework-camel-case==1.0.3
drf-yasg==1.15.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
phonenumbers==8.10.9
pytz==2019.1
requests==2.21.0
ruamel.yaml==0.15.92
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
